cart impl and controller

+++ src/main/java/com/example/demo/services/CartProductService.java
@@ -5,8 +5,8 @@ import com.example.demo.model.CartProduct;
 import com.example.demo.model.Product;
 
 public interface CartProductService {
-    CartProduct createCartProduct(int quantity, int totalPrice, Cart cart, Product product);
-    CartProduct updateCartProduct(Long id, int quantity, int totalPrice, Cart cart, Product product);
+    CartProduct createCartProduct(int quantity, Float totalPrice, Cart cart, Product product);
+    CartProduct updateCartProduct(Long id, int quantity, Float totalPrice, Cart cart, Product product);
     CartProduct deleteCartProduct(Long id);
     CartProduct getCartProduct(Long id);
 }
